env_demo/env_1533_initv-data-pods/pyproject.toml
[project]
name = "env_1533_initv-data-pods"
version = "0.1.0"
description = "Data Pods - Create and manage modular portable database pods"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "fastmcp>=1.0",
    "httpx>=0.27.0",
    "mcp[server]==1.1.2",
    "pydantic>=2.7.2",
    "requests==2.31.0",
    "uvicorn[standard]==0.27.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
]